% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{convergence_check}
\alias{convergence_check}
\title{Check stan model fit for convergence (not yet implemented)}
\usage{
convergence_check(fit, ...)
}
\arguments{
\item{fit}{unused}

\item{...}{unused}
}
\description{
Nothing here yet
}
\details{
Nothing here yet
}
\examples{
runif(1)
}
