% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{data_analyst}
\alias{data_analyst}
\title{Fit a Stan/JAGS model in a single data set}
\usage{
data_analyst(i, raw = raw, fl = "~/temp.csv", s.code = NULL,
  s.file = NULL, iter = 2500, warmup = 500, chains = 4, p = NULL,
  dx = NULL, N = NULL, verbose = FALSE, type = "stan",
  stanfit = NA, ...)
}
\arguments{
\item{i}{iteration in simulated data}

\item{raw}{data frame with raw data}

\item{fl}{Filename to output model results for single run (useful for debugging)}

\item{s.code}{name of a character string with a stan model}

\item{s.file}{(experimental) character name of a presepecified stan model ("wwbd_simtemplate_20190205", "wwbd_simlogistic_TransParameter_0211201920190211")}

\item{iter}{Number of "sweep" iterations, or iterations after the burnin/warmup}

\item{warmup}{warmup (stan) / n.adapt (jags). Number of iterations to allow
for adaptation of MCMC parameters/burnin}

\item{chains}{Number of parallel MCMC chains (default 4)}

\item{p}{number of beta parameters in outcome model (optional - R makes a guess)}

\item{dx}{number of columns in data matrix (optional - R makes a guess)}

\item{N}{Sample size (optional - R makes a guess)}

\item{verbose}{print extra debugging info? FALSE}

\item{type}{Which type of model is it? Can be 'stan' or 'jags'}

\item{stanfit}{Name of stan_model or stan output that can be recycled. This
will use a pre-compiled version of the stan code which cuts simulation time
significantly over multiple runs.}

\item{...}{arguments to stan() or coda.samples() for a Stan or JAGS model, 
respectively}
}
\description{
Fit a Stan/JAGS model in a single data set
}
\details{
Fit a Stan/JAGS model in a single data set
}
\examples{
runif(1)
}
