% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{data_reader}
\alias{data_reader}
\title{lorem ipsum}
\usage{
data_reader(raw, i, expnm = c("Arsenic", "Manganese", "Lead", "Cadmium",
  "Copper"), p = 10, dx = 5, N = NULL, verbose = FALSE, ...)
}
\arguments{
\item{raw}{data frame with raw data}

\item{i}{iteration of raw data to read}

\item{expnm}{character vector of exposure names}

\item{p}{number of beta parameters in outcome model (optional - R makes a guess)}

\item{dx}{number of columns in data matrix (optional - R makes a guess)}

\item{N}{Sample size (optional - R makes a guess)}

\item{verbose}{print extra debugging info? FALSE}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
