% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{get_model}
\alias{get_model}
\title{Get an existing model from the wellwise package}
\usage{
get_model(filename = "logistic.stan", ...)
}
\arguments{
\item{filename}{character string with stub, or name (stub.stan) of file in /inst directory that contains a stan model}

\item{...}{UNUSED}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
