% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_fullrmod}
\alias{make_fullrmod}
\title{make a model to use in model.matrix}
\usage{
make_fullrmod(terms, intvars, binvars = NULL)
}
\arguments{
\item{terms}{ipsum}

\item{intvars}{ipsum}

\item{binvars}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
