% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkintervention}
\alias{mkintervention}
\title{lorem ipsum}
\usage{
mkintervention(mod, vars, subs)
}
\arguments{
\item{mod}{ipsum}

\item{vars}{ipsum}

\item{subs}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
