% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkinterventionR}
\alias{mkinterventionR}
\title{helper: create R code that encodes an intervention in g-computation}
\usage{
mkinterventionR(mod, vars, subs)
}
\arguments{
\item{mod}{existing model code}

\item{vars}{variable names to intervene on}

\item{subs}{intervention values}
}
\description{
not used by user, typically
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
