% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkints}
\alias{mkints}
\title{helper: lorem ipsum}
\usage{
mkints(terms, intvars, binvars)
}
\arguments{
\item{terms}{ipsum}

\item{intvars}{ipsum}

\item{binvars}{binary variables (character vector) which are excluded to avoid meaningless self-interactions}
}
\description{
not used by user, typically
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
