% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkmod}
\alias{mkmod}
\title{helper: create a model based on generic terms}
\usage{
mkmod(coef, terms, start = 1, catmod = FALSE)
}
\arguments{
\item{coef}{ipsum}

\item{terms}{ipsum}

\item{start}{ipsum}

\item{catmod}{ipsum}
}
\description{
not used by user, typically
}
\details{
(mkmodidx probably better)
}
\examples{
runif(1)
}
