% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkmodidx}
\alias{mkmodidx}
\title{helper: create a model based on generic terms,}
\usage{
mkmodidx(coef, terms, start = 1, index = "i", print = FALSE,
  indexed = FALSE)
}
\arguments{
\item{coef}{ipsum}

\item{terms}{ipsum}

\item{start}{ipsum}

\item{index}{ipsum}

\item{print}{ipsum}

\item{indexed}{ipsum}
}
\description{
not used by user, typically
}
\details{
including interactions, and output model as a string - preferred way
 to make stan model code
}
\examples{
runif(1)
}
