% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{print}
\alias{print}
\alias{print.bgfsimres}
\title{lorem ipsum}
\usage{
\method{print}{bgfsimres}(x, ...)
}
\arguments{
\item{x}{ipsum}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
