% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{sink.reset}
\alias{sink.reset}
\title{clean up problems with sinking where output doesn't show up}
\usage{
sink.reset(...)
}
\arguments{
\item{...}{unused}
}
\description{
analytic functions in wellwise use `sink` to avoid 
polluting the output with progress bars. A side effect is that sometimes
output can be turned off altogether. Runk sink.reset() after data_analyst
as a way to fix this problem
}
\details{
lorem ipsum
}
\examples{
sink.reset()    
}
