% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sortstring}
\alias{sortstring}
\title{helper: Sort the words in a character string}
\usage{
sortstring(s)
}
\arguments{
\item{s}{ipsum}
}
\description{
not used by user, typically
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
