% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{summary.mcmc.list}
\alias{summary.mcmc.list}
\title{Summary method for mcmc.list object from coda package}
\usage{
\method{summary}{mcmc.list}(...)
}
\arguments{
\item{...}{arguments to coda:::summary.mcmc.list (object, quantiles=c(.025, .25, .5, .75, .975))}
}
\description{
This function was taken directly from the `coda` package. Not
exported in that package.
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
