% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trimints}
\alias{trimints}
\title{helper: lorem ipsum}
\usage{
trimints(trimvars, terms)
}
\arguments{
\item{trimvars}{ipsum}

\item{terms}{ipsum}
}
\description{
not used by user, typically
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
