% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{welldata}
\alias{welldata}
\title{Well water measurements}
\format{A data frame with 52440 rows and 29 variables:
\describe{
  \item{Arsenic}{Arsenic measurement}
  \item{Barium}{Barium measurement}
  \item{Cadmium}{Cadmium measurement}
  \item{Calcium}{Calcium measurement}
  \item{Chloride}{Chloride measurement}
  \item{Chromium}{Chromium measurement}
  \item{Copper}{Copper measurement}
  \item{Fluoride}{Fluoride measurement}
  \item{Hexavalent_Chromium}{Hexavalent Chromium measurement}
  \item{Insoluble_Iron}{Insoluble Iron measurement}
  \item{Iron}{Iron measurement}
  \item{Lead}{Lead measurement}
  \item{Magnesium}{Magnesium measurement}
  \item{Manganese}{Manganese measurement}
  \item{Mercury}{Mercury measurement}
  \item{Nitrate}{Nitrate measurement}
  \item{Nitrite}{Nitrite measurement}
  \item{pH}{pH measurement}
  \item{Selenium}{Selenium measurement}
  \item{Silver}{Silver measurement}
  \item{Sodium}{Sodium measurement}
  \item{Soluble_Iron}{Soluble Iron measurement}
  \item{Sulfate}{Sulfate measurement}
  \item{Total_Alkalinity}{Total Alkalinity measurement}
  \item{Total_Hardness}{Total Hardness measurement}
  \item{Uranium}{Uranium measurement}
  \item{Zinc}{Zinc measurement}
  \item{iter}{Iteration number for simulation}
  \item{y}{Simulated outcome}
}}
\source{
\url{https://celr.ncpublichealth.com}
}
\usage{
welldata
}
\description{
A dataset containing well water measurements and simulated
outcomes for 52440 units across 20 iterations
}
\keyword{datasets}
