% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{analysis_wrapper}
\alias{analysis_wrapper}
\title{lorem ipsum}
\usage{
analysis_wrapper(iter, rawdata, dir = "~/temp/", root, debug = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{iter}{ipsum}

\item{rawdata}{ipsum}

\item{dir}{ipsum}

\item{root}{ipsum}

\item{debug}{ipsum}

\item{verbose}{ipsum}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
