% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{analysis_wrapper}
\alias{analysis_wrapper}
\title{lorem ipsum}
\usage{
analysis_wrapper(simiters, rawdata, dir = "~/temp/", root,
  debug = FALSE, verbose = FALSE, type = "stan", stanfit = NA, ...)
}
\arguments{
\item{simiters}{1:2}

\item{rawdata}{RAW}

\item{dir}{"~/temp/"}

\item{root}{"test"}

\item{debug}{FALSE}

\item{verbose}{FALSE}

\item{type}{'jags' or 'stan'}

\item{stanfit}{NA or name of object containing a fitted stan model (re-uses
the compiled stan model, which will be efficient for restarting iterations after
checking first one(s) for diagnostics}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
