% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{convergence_check}
\alias{convergence_check}
\title{lorem ipsum}
\usage{
convergence_check(fit, ...)
}
\arguments{
\item{fit}{ipsum}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
