% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{data_analyst}
\alias{data_analyst}
\title{lorem ipsum}
\usage{
data_analyst(i, raw = raw, fl = "~/temp.csv", s.code = NULL,
  iter = 2500, warmup = 500, p = NULL, dx = NULL, N = NULL, ...)
}
\arguments{
\item{i}{ipsum}

\item{raw}{ipsum}

\item{fl}{ipsum}

\item{s.code}{ipsum}

\item{iter}{ipsum}

\item{warmup}{ipsum}

\item{p}{ipsum}

\item{dx}{ipsum}

\item{N}{ipsum}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
