% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{data_analyst}
\alias{data_analyst}
\title{lorem ipsum}
\usage{
data_analyst(i, raw = raw, fl = "~/temp.csv", s.code = NULL,
  s.file = NULL, iter = 2500, warmup = 500, chains = 4, p = NULL,
  dx = NULL, N = NULL, verbose = FALSE, type = "stan", ...)
}
\arguments{
\item{i}{ipsum}

\item{raw}{ipsum}

\item{fl}{ipsum}

\item{s.code}{name of a character string with a stan model}

\item{s.file}{character name of a presepecified stan model ("wwbd_simtemplate_20190205", "wwbd_simlogistic_TransParameter_0211201920190211")}

\item{iter}{ipsum}

\item{warmup}{ipsum}

\item{chains}{ipsum}

\item{p}{ipsum}

\item{dx}{ipsum}

\item{N}{ipsum}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
