% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{data_importer}
\alias{data_importer}
\title{lorem ipsum}
\usage{
data_importer(package = NULL, ...)
}
\arguments{
\item{package}{import data from a specific package (NULL or 'wellwise')}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
