% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{data_reader}
\alias{data_reader}
\title{lorem ipsum}
\usage{
data_reader(raw, i, expnm = c("Arsenic", "Manganese", "Lead", "Cadmium",
  "Copper"), p = 10, dx = 5, N = NULL, ...)
}
\arguments{
\item{raw}{ipsum}

\item{i}{ipsum}

\item{expnm}{ipsum}

\item{p}{ipsum}

\item{dx}{ipsum}

\item{N}{ipsum}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
