% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{julia.model}
\alias{julia.model}
\title{Fit a julia gibbs sampler (given by s.code)}
\usage{
julia.model(iter, j.code, warmup, chains, sdat = sdat, fl = tempfile(),
  jinstance = NULL, cleanafter = FALSE, verbose = TRUE,
  debug = FALSE)
}
\arguments{
\item{iter}{Number of "sweep" iterations, or iterations after the burnin/warmup}

\item{j.code}{name of a character string with a stan model}

\item{warmup}{warmup (stan) / n.adapt (jags) / burnin (julia). Number of iterations to allow
for adaptation of MCMC parameters/burnin}

\item{chains}{Number of parallel MCMC chains (default 4)}

\item{sdat}{list of data created from}

\item{fl}{Filename to output model results for single run (useful for debugging)}

\item{jinstance}{existing julia_setup() instance}

\item{cleanafter}{should the extra workers be killed after running?}

\item{verbose}{print extra stuff}

\item{debug}{print extra debugging info? FALSE}
}
\description{
Fit a julia model to a single data set
}
\details{
Fit a julia model to a single data set
}
