% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_fullrmod}
\alias{make_fullrmod}
\title{helper: make a model to use in model.matrix}
\usage{
make_fullrmod(terms, intvars, binvars = NULL)
}
\arguments{
\item{terms}{ipsum}

\item{intvars}{ipsum}

\item{binvars}{ipsum}
}
\description{
not used by user, typically
}
\details{
Creates an R formula given a set of terms and interaction variables
}
\examples{
runif(1)
}
