% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkdesignR}
\alias{mkdesignR}
\title{helper: Create code for a design matrix}
\usage{
mkdesignR(terms, catmod = FALSE)
}
\arguments{
\item{terms}{ipsum}

\item{catmod}{ipsum}
}
\description{
not used by user, typically
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
