% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkints}
\alias{mkints}
\title{lorem ipsum}
\usage{
mkints(terms, intvars)
}
\arguments{
\item{terms}{ipsum}

\item{intvars}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
