% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mkmod}
\alias{mkmod}
\title{lorem ipsum}
\usage{
mkmod(coef, terms, start = 1, catmod = FALSE)
}
\arguments{
\item{coef}{ipsum}

\item{terms}{ipsum}

\item{start}{ipsum}

\item{catmod}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
