% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{plot}
\alias{plot}
\alias{plot.bgfsimres}
\title{lorem ipsum}
\usage{
\method{plot}{bgfsimres}(x, type = ifelse(nrow(x$postmeans) > 50,
  "density", "hist"), ...)
}
\arguments{
\item{x}{ipsum}

\item{type}{"density" "hist"}

\item{...}{ipsum}
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)

}
