% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{summary}
\alias{summary}
\alias{summary.bgfsimmodlist}
\title{Summarize results of multiple simulations}
\usage{
\method{summary}{bgfsimmodlist}(object = NULL, ...)
}
\arguments{
\item{object}{ipsum}

\item{...}{ipsum}
}
\value{
a 'bgfsimres' object
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
\examples{
runif(1)
}
